+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Threading/ObjectHeader.cs
@@ -239,7 +239,7 @@ namespace System.Threading
                     SyncTable.MoveThinLockToNewEntry(
                         syncIndex,
                         oldBits & SBLK_MASK_LOCK_THREADID,
-                        (ushort)((oldBits & SBLK_MASK_LOCK_RECLEVEL) >> SBLK_RECLEVEL_SHIFT));
+                        (uint)((oldBits & SBLK_MASK_LOCK_RECLEVEL) >> SBLK_RECLEVEL_SHIFT));
                 }
 
                 // Store the sync entry index
